python/foundryHA/requirements.txt
azure-ai-agentserver-agentframework==1.0.0b16
python-dotenv
azure-identity
a2a-sdk>=0.3.20
httpx>=0.27.0
colorama>=0.4.6
debugpy
agent-dev-cli
